requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
pandas
langchain
numpy
python-dotenv
gradio